multiple drop points and amount precision added

+++ lib/models/direction_response_model.dart
@@ -1,6 +1,6 @@
 class DirectionDetailsInfo{
-  int? distance_value;
-  int? duration_value;
+  double? distance_value;
+  double? duration_value;
   String? e_points;
   String? distance_text;
   String? duration_text;
